Upgraded from 1.2.0 to 1.1.5 Fork from github and push back to github
Changes to Guided Tour for WordPress.com
- cookie: guiders property allows you to name a cookie that gets
updated every time show() is called. Requires jQuery Cookies plugin
(https://github.com/carhartl/jquery-cookie)
- failStep: guiders property allows you to name a step to show() if the
show() case fails (attachTo element is missing). For obvious reasons,
this should not have an attachTo
- resume(): start up tour from current place in cookie (if set). This
is useful when your tour leaves the page you are on. Unlike show, it
will skip steps that need to be skipped.
- endTour(): Like hideAll() but it remembers to remove the cookie
position.
 - initGuider(): Allows for initializing Guiders without actually
creating them (useful when guider is not in the DOM yet. Avoids error:
base is null [Break On This Error] var top = base.top;
- autoAdvance: property allows binding to an element (and event) to
auto-advance the guider. This is a combination of onShow() binding plus
removing of bind when next is done.
- shouldSkip: property defines a function handler forces a skip of this
step if function returns true.
- overlay: If not set to true, this defines the class of the overlay.
(This is useful for coloring the background of the overlay red on error
with a class="error")