Since we de-highlight guider.highlight, we should be highlighting it instead of guider.attachTo.
While it is common to set highlight to the same ID as attachTo, I often attach to one thing but highlight another--like the attachTo's container. Also, must the guider have an attachTo for a highlight to be valid? If not, the test of attachTo is also unnecessary. It's no harm right now, so I left it in.